F01: Add implementation of RegionService
